Update Node binary SHA sum
In #83 we updated the version of Node to take some security fixes,
however we forgot to also update the SHA sum we chack the downloaded
binary against. Update this to the correct string.

There's no need to update the version tag - we've not been able to
build and release 10.0.2 yet.